% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.copy}
\alias{bursts.copy}
\title{Copy a list of bursts (by value)}
\usage{
bursts.copy(bursts)
}
\arguments{
\item{bursts}{bursts to copy}
}
\value{
A copy of the bursts.
}
\description{
Copy a list of bursts (by value)
}
