% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.filter}
\alias{bursts.filter}
\title{From a list of bursts, extract those that interest you by
passing a selecting function. See the examples.}
\usage{
bursts.filter(bursts, func, one_file = FALSE)
}
\arguments{
\item{bursts}{The list of all bursts}

\item{func}{A function of a record that returns either TRUE or FALSE}

\item{one_file}{TRUE or FALSE: Return a single file to disk, or a list of bursts.
The one_file will return a file with all filtered bursts zeroed out.}
}
\value{
A shorter list of bursts OR if one_file is passed one record with zeros where the other bursts might have been originally. Defaults to FALSE.
}
\description{
From a list of bursts, extract those that interest you by
passing a selecting function. See the examples.
}
\examples{
\dontrun{
high_popen <- function (rec) {

    record.popen(rec) > 0.7

}

subset <- bursts.filter(high_popen, bursts)


# To export to one .dwt file
subset_f <- bursts.filter(high_popen, bursts, one_file=TRUE)
}
}
