% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.get_breaks}
\alias{bursts.get_breaks}
\title{Extract vector of breaks from the chunks}
\usage{
bursts.get_breaks(chunks)
}
\arguments{
\item{chunks}{The list of segments}
}
\value{
A vector of break times
}
\description{
This is done using the start_time attribute, which
is mostly hidden in the data.
}
\examples{

breaks <- bursts.get_breaks(chunks)

}
