% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.pcloseds}
\alias{bursts.pcloseds}
\title{Return pcloseds of every burst.}
\usage{
bursts.pcloseds(bursts)
}
\arguments{
\item{bursts}{The list of all bursts}
}
\value{
The pclosed values
}
\description{
Return pcloseds of every burst.
}
\examples{
\dontrun{
pcloseds <- bursts.popens(bursts)
hist(pcloseds)
}
}
