% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.popens}
\alias{bursts.popens}
\title{Return popens of every burst.}
\usage{
bursts.popens(bursts)
}
\arguments{
\item{bursts}{The list of all bursts}
}
\value{
The popen values
}
\description{
Return popens of every burst.
}
\examples{

infile <- system.file("extdata", "example_corrected.dwt", package = "scbursts")
dwells_c <- dwt.read(infile)
bursts <- bursts.defined_by_tcrit(dwells_c, 1.511842, units="s")

popens <- bursts.popens(bursts)
hist(popens)

}
