% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.remove_first_and_last}
\alias{bursts.remove_first_and_last}
\title{Remove the first and last burst from the list}
\usage{
bursts.remove_first_and_last(chunks)
}
\arguments{
\item{chunks}{The list of all bursts}
}
\value{
A shorter list of bursts
}
\description{
Remove the first and last burst from the list
}
\examples{

chunks <- bursts.remove_first_and_last(chunks)

}
