% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.sort}
\alias{bursts.sort}
\title{Order a list of bursts by some function. For instance, popen.}
\usage{
bursts.sort(chunks, func, reverse = FALSE)
}
\arguments{
\item{chunks}{The list of all bursts}

\item{func}{A function of a segment that returns a numeric value}

\item{reverse}{By default, return in ascending order. Use reverse=TRUE to change that.}
}
\value{
A list sorted by func. By default in ascending order (unless reversed)
}
\description{
Order a list of bursts by some function. For instance, popen.
}
\examples{
\dontrun{
sorted <- bursts.sort(segment.popen, chunks)
}
}
