% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.start_times_update}
\alias{bursts.start_times_update}
\title{Attach the meta-data to each record saying when it began.}
\usage{
bursts.start_times_update(bursts, gaps)
}
\arguments{
\item{bursts}{List of records}

\item{gaps}{vector of gap times.}
}
\value{
A list of records, one per burst, with updated start_times
}
\description{
It interleaves the durations of the bursts and gaps, and
assigns the sum of those durations up to a point as the
starting time
}
\details{
You probably won't ever have to call this directly.
}
