% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk_bursts}
\alias{chunk_bursts}
\title{Split record at long pauses, dividing the record
into multiple -shorter- bursts.}
\usage{
chunk_bursts(table, t_crit)
}
\arguments{
\item{table}{Table with $states and $dwells}

\item{t_crit}{Critical time (us) at which to divide bursts}
}
\value{
list of columns of dataframes (one per row)
}
\description{
Split record at long pauses, dividing the record
into multiple -shorter- bursts.
}
\examples{
chunks <- chunk_bursts(table, 14.77155587)
}
