% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risetime.R
\name{correct_risetime}
\alias{correct_risetime}
\title{Undo the effect of the guassian filter
See section 4.1.1 of Colquhoun and Sigworth,
"Fitting and Analysis of Single-Channel records"}
\usage{
correct_risetime(Tr, segment)
}
\arguments{
\item{Tr}{Rise time of the filter in (us)}

\item{segment}{segment with $states and $dwells to correct.}
}
\value{
Segment with corrected risetimes.
}
\description{
NOTE: This is potentially problematic, in that this unfiltering
lengthens every state. A less naïve algorithm would take into
account the infulence of the surroundings, as they impact the
effects of the filter.
}
\examples{
segment <- correct_risetime(14.77155587, segment)
}
