% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot.log_root_axes}
\alias{cplot.log_root_axes}
\title{Add log-root axes to histogram plot}
\usage{
cplot.log_root_axes(points)
}
\arguments{
\item{points}{The data to plot}
}
\description{
Add log-root axes to histogram plot
}
\examples{

infile <- system.file("extdata", "example_corrected.dwt", package = "scbursts")
dwells_c <- dwt.read(infile)
bursts <- bursts.defined_by_tcrit(dwells_c, 1.511842, units="s")

open_dwells <- segment.open_dwells(bursts.recombine(bursts))
hist(log10(open_dwells), axes=FALSE, breaks=30)
cplot.log_root_axes(open_dwells)

}
