% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.log_root_hist}
\alias{cplot.log_root_hist}
\title{Plot dwells as a histogram after log transforming}
\usage{
cplot.log_root_hist(dwells, ...)
}
\arguments{
\item{dwells}{The data to plot}

\item{...}{Any and options you can pass to plot, except for}

\item{title}{The title of the plot. Defaults to blank}
}
\description{
Plot dwells as a histogram after log transforming
}
\examples{

\dontrun{
open_times   <- subset(table, states == 1)$dwells
closed_times <- subset(table, states == 0)$dwells

cplot.log_root_hist(open_times, "Open Times")
cplot.log_root_hist(closed_times, "Closed Times")
}

}
\seealso{
plot
}
