% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.pclosed_hist}
\alias{cplot.pclosed_hist}
\title{Plot Histogram of P(Closed)}
\usage{
cplot.pclosed_hist(bursts, title = "P(Closed) Histogram",
  y_transform = identity)
}
\arguments{
\item{bursts}{List of multiple records}

\item{title}{The title of the plot.}

\item{y_transform}{Transform the y_axis (for instance, log10 it)}
}
\description{
Plot Histogram of P(Closed)
}
\examples{
\dontrun{
cplot.pclosed_hist(bursts, "P(Closed) Histogram, 2017-09-14")
cplot.pclosed_hist(bursts, "P(Closed) Histogram, 2017-09-14", y_transform=log10)
}
}
