% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot.pclosed_ts}
\alias{cplot.pclosed_ts}
\title{Plot Time Series (ts) of P(Closed).}
\usage{
cplot.pclosed_ts(bursts, main = "P(Closed) Time Series", xlim = NULL)
}
\arguments{
\item{bursts}{List of multiple segments}

\item{main}{The title of the plot.}

\item{xlim}{Pass xlim argument to plot() to focus on window of time series}
}
\description{
Plot Time Series (ts) of P(Closed).
}
\examples{

infile <- system.file("extdata", "example_corrected.dwt", package = "scbursts")
dwells_c <- dwt.read(infile)
bursts <- bursts.defined_by_tcrit(dwells_c, 1.511842, units="s")

cplot.pclosed_ts(bursts, "P(Closed) Time Series, 2018-09-20")

}
