% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.popen_hist}
\alias{cplot.popen_hist}
\title{Plot Histogram of P(Open)}
\usage{
cplot.popen_hist(chunks, title = "P(Open) Histogram",
  y_transform = identity)
}
\arguments{
\item{chunks}{List of multiple segments}

\item{title}{The title of the plot.}

\item{y_transform}{Transform the y_axis (for instance, log10 it)}
}
\description{
Plot Histogram of P(Open)
}
\examples{

cplot.popen_hist(chunks, "P(Open) Histogram, 2017-09-14")
cplot.popen_hist(chunks, "P(Open) Histogram, 2017-09-14", y_transform=log10)

}
