% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.popen_ts}
\alias{cplot.popen_ts}
\title{Plot Time Series of P(Open)}
\usage{
cplot.popen_ts(chunks, title = "P(Open) Time Series")
}
\arguments{
\item{chunks}{List of multiple segments}

\item{title}{The title of the plot.}
}
\description{
Plot Time Series of P(Open)
}
\examples{

cplot.popen_ts(chunks, "P(Open) Time Series, 2017-09-14")

}
