% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.read}
\alias{dwt.read}
\title{Read a .dwt file. Result is a "segment", which is a
dataframe extra data. See "segment" for more details.}
\usage{
dwt.read(filename)
}
\arguments{
\item{filename}{Filename to read from}
}
\description{
Read a .dwt file. Result is a "segment", which is a
dataframe extra data. See "segment" for more details.
}
\examples{
\dontrun{
seg <- dwt.read(segment, "file-test.dwt")
}
}
