% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.read_bursts}
\alias{dwt.read_bursts}
\title{Read chunks from a folder.}
\usage{
dwt.read_bursts(folder)
}
\arguments{
\item{folder}{list of dataframes corresponding to bursts}
}
\value{
A pair (chunks,breaks), where the chunks are segments
starting and ending in 1 states, and breaks is a vector of 0s which sit
inbetween the bursts. There will be n chunks and n-1 breaks.
}
\description{
Read chunks from a folder.
}
\examples{
\dontrun{
pair <- dwt.read_bursts("bursts/60uM-2017-08-19-19-35")
chunks <- pair$chunks
breaks <- pair$breaks
}
}
