% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.write}
\alias{dwt.write}
\title{Reformat the digits of data before writing to disk.}
\usage{
dwt.write(segment, file = "", seg = 1)
}
\arguments{
\item{segment}{segment with $dwells and $states}

\item{file}{Filename to write to}

\item{seg}{Segment to write in .dwt header.}
}
\description{
Writes DOS line endings.
}
\examples{
\dontrun{
dwt.write(segment, "file-test.dwt")
}
}
