% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.write_bursts}
\alias{dwt.write_bursts}
\title{Write all the bursts to dwt. They will be placed in
a timestamped subfolder of bursts/}
\usage{
dwt.write_bursts(chunks, breaks = c(), directory = "bursts",
  filename = "burst", timestamp = TRUE)
}
\arguments{
\item{chunks}{list of dataframes corresponding to bursts}
}
\value{
Name of folder containing contents.
}
\description{
All bursts will be written out as .dwt files, labeled
in temporal order.
}
\examples{
dwt.write_chunks(chunks)
}
