% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.extract_header}
\alias{evt.extract_header}
\title{Extract header from evt file}
\usage{
evt.extract_header(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
A string containing the header
}
\description{
Extract header from evt file
}
\examples{
\dontrun{
header <- evt.extract_header("data/60uM.evt")

evt.write(segments, header=header, file="60uMc.evt")

}
}
