% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.from_dwells}
\alias{evt.from_dwells}
\title{Converts dwell durations to absolute transition times}
\usage{
evt.from_dwells(segment)
}
\arguments{
\item{segment}{A segment}
}
\value{
A dataframe of states and transition times
}
\description{
Converts dwell durations to absolute transition times
}
\examples{

\dontrun{
table <- evt.from_dwells(segment)
}
}
