% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.read}
\alias{evt.read}
\title{Read a .evt file to a table}
\usage{
evt.read(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
A table with columns "states" and "times"
}
\description{
Read a .evt file to a table
}
\examples{
table <- evt.read("data/60uM.evt")

# import some of the data included with the package
infile <- system.file("extdata", "60uM.evt", package = "uottawaionchannel")
table <- evt.read(infile)

}
