% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.to_dwells}
\alias{evt.to_dwells}
\title{Calculate pulse lengths}
\usage{
evt.to_dwells(table)
}
\arguments{
\item{table}{with columns "states" and "times"}
}
\value{
A "record" with one less row, where each row

states    dwells
0         0.016010

represents pulse in state 0 of duration 0.51231, instead
of the time at which the state transitioned.

Also, the output table preserves the attributes

See "record" for more info
}
\description{
Converts transition times to dwell lengths
}
\examples{
\dontrun{
record <- evt.to_dwells(table)
}
}
