% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.write}
\alias{evt.write}
\title{Write burst(s) to a .evt file}
\usage{
evt.write(segment, file = "")
}
\arguments{
\item{segment}{A single segment to write to filename}

\item{filename}{The filename}
}
\description{
Write burst(s) to a .evt file
}
\examples{
\dontrun{

evt.write(segment, file="60mc-2.evt")

}
}
