% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hst.R
\name{hst.extract_header}
\alias{hst.extract_header}
\title{Extract header from hst file}
\usage{
hst.extract_header(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
A string containing the header
}
\description{
Extract header from hst file
}
\examples{
\dontrun{
header <- hst.extract_header("data/60uM.hst")

hst.write(open_table, closed_table, file="60uMc.hst", header=header)

}
}
