% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hst.R
\name{hst.read}
\alias{hst.read}
\title{Read a .hst file to a table (x, y, fit)}
\usage{
hst.read(filename, extract = "open", square = FALSE)
}
\arguments{
\item{filename}{The filename}

\item{extract}{Extract either "open" or "closed" histogram}

\item{square}{Reverse a square-root operation on histogram}
}
\value{
A tables with columns "x", "y" and "fit".
}
\description{
Times are in seconds
}
\examples{

library(uottawaionchannel)

\dontrun{
mil_hst <- hst.read("data/A1W18_dur.hst", extract="open", square=TRUE)
}

# import some of the data included with the package
infile <- system.file("extdata", "example.hst", package = "uottawaionchannel")
table <- hst.read(infile)

table

}
