% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hst.R
\name{hst.write}
\alias{hst.write}
\title{Write bursts to a .hst file}
\usage{
hst.write(open_hist, closed_hist, file = "", header = NULL)
}
\arguments{
\item{segments}{A segment or list of segments to write to filename}

\item{filename}{The filename}
}
\description{
Write bursts to a .hst file
}
\examples{
\dontrun{

hst.write(, file="60uMc-R.hst")

}
}
