% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{popen}
\alias{popen}
\title{Calculate empirical P(Open) of a segment}
\usage{
popen(segment)
}
\arguments{
\item{segment}{The dwells and states table}
}
\value{
The ratio of open time to total time
}
\description{
NOTE: Removing first and last state, as they are pauses.
}
\examples{
popen(segment)
> 0.22
}
