% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{read.dwt}
\alias{read.dwt}
\title{Read a .dwt file. Result is a "segment", which is a
dataframe extra data. See "segment" for more details.}
\usage{
read.dwt(filename)
}
\arguments{
\item{filename}{Filename to read from}
}
\description{
Read a .dwt file. Result is a "segment", which is a
dataframe extra data. See "segment" for more details.
}
\examples{
seg <- read.dwt(segment, "file-test.dwt")
}
