% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{read.evt}
\alias{read.evt}
\title{Read a .evt file to a table}
\usage{
read.evt(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
A table with columns "states" and "times"
}
\description{
Read a .evt file to a table
}
\examples{
table <- read.evt("data/60uM.evt")

# import some of the data included with the package
infile <- system.file("extdata", "60uM.evt", package = "uottawaionchannel")
table <- read.evt(infile)

}
