% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.count_closed}
\alias{record.count_closed}
\title{Extract number of closed dwells}
\usage{
record.count_closed(record)
}
\arguments{
\item{record}{the record object}
}
\value{
number of closed dwells
}
\description{
Extract number of closed dwells
}
\examples{
\dontrun{
record.count_closed(data)
> 5
}
}
