% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.count_open}
\alias{record.count_open}
\title{Extract number of open dwells}
\usage{
record.count_open(record)
}
\arguments{
\item{record}{the record object}
}
\value{
number of open dwells
}
\description{
Extract number of open dwells
}
\examples{
\dontrun{
record.count_open(data)
> 5
}
}
