% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.create}
\alias{record.create}
\title{Create object containing table data and metadata}
\usage{
record.create(states, dwells, rec = 1, start_time = 0, name = "burst")
}
\arguments{
\item{states}{a vector of states in {0,1}}

\item{dwells}{a vector of dwell durations (same length as states)}

\item{rec}{The record number. Defaults to 1}

\item{start_time}{When the dwells began. Defaults to 0}

\item{name}{Suffix-less version of the original filename. 60uM.dwt -> '60uM'}
}
\value{
The record object: A dataframe with extra metadata.
}
\description{
The object can be used as a dataframe, and the
metadata can be accessed with the functions
}
\details{
- record.rec
- record.start_time
- record.filename
}
\examples{
\dontrun{
record <- record.create(states, dwells, rec=12, start_time=0, name="60uMc")
record.name(record)
> "60uMc"
}
}
