% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.duration}
\alias{record.duration}
\title{Get duration of a record}
\usage{
record.duration(record)
}
\arguments{
\item{record}{the record object}
}
\value{
the duration
}
\description{
Get duration of a record
}
\examples{
\dontrun{
record.duration(data)
> 2.11
}
}
