% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.name}
\alias{record.name}
\title{Extract name from record}
\usage{
record.name(record)
}
\arguments{
\item{record}{the record object}
}
\value{
Record name (string)
}
\description{
Extract name from record
}
\examples{
\dontrun{
record.name(data)
> "60uMc"
}
}
