% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.pclosed}
\alias{record.pclosed}
\title{Calculate empirical P(Closed) of a record}
\usage{
record.pclosed(record)
}
\arguments{
\item{record}{The dwells and states table}
}
\value{
The ratio of closed time to total time
}
\description{
NOTE: Assuming that burst starts and ends with 1
}
\examples{
\dontrun{
record.pclosed(record)
> 0.78
}
}
