% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.popen}
\alias{record.popen}
\title{Calculate empirical P(Open) of a record}
\usage{
record.popen(record)
}
\arguments{
\item{record}{The dwells and states table}
}
\value{
The ratio of open time to total time
}
\description{
NOTE: Assuming that burst starts and ends with 1
}
\examples{
\dontrun{
record.popen(record)
> 0.22
}
}
