% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.rec}
\alias{record.rec}
\title{Extract record number from record}
\usage{
record.rec(record)
}
\arguments{
\item{record}{the record object}
}
\value{
Record number (integer)
}
\description{
Extract record number from record
}
\examples{
\dontrun{
record.rec(data)
> 12
}
}
