% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.start_time}
\alias{record.start_time}
\title{Extract start_time from record}
\usage{
record.start_time(record)
}
\arguments{
\item{record}{the record object}
}
\value{
Record start_time (float)
}
\description{
Extract start_time from record
}
\examples{
\dontrun{
record.start_time(data)
> 17.123295
}
}
