% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{relative_time}
\alias{relative_time}
\title{Calculate pulse lengths}
\usage{
relative_time(table)
}
\arguments{
\item{table}{with columns "states" and "times"}
}
\value{
A table with one less row, where each row

states    dwells
0         0.016010

represents pulse in state 0 of duration 0.51231, instead
of the time at which the state transitioned.
}
\description{
!!! TIMES becomes DWELLS !!!
}
\examples{
table <- relative_time(table) # updated
}
