% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risetime.R
\name{risetime.correct_guassian}
\alias{risetime.correct_guassian}
\title{Undo the effect of the guassian filter
See section 4.1.1 of Colquhoun and Sigworth,
"Fitting and Analysis of Single-Channel records"}
\usage{
risetime.correct_guassian(Tr, segment)
}
\arguments{
\item{Tr}{Rise time of the filter in (us)}

\item{segment}{segment with $states and $dwells to correct.}
}
\value{
Segment with corrected risetimes.
}
\description{
NOTE: This is potentially problematic, in that this unfiltering
lengthens every state. A less naive algorithm would take into
account the infulence of the surroundings, as they impact the
effects of the filter.
}
\examples{
\dontrun{
segment <- risetime.correct_guassian(14.77155587, segment)
}
}
