% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan.read}
\alias{scan.read}
\title{Read a scan results text file. scan.read returns a 1 segment list
Reads in scan results and puts them in the same format as the output
of dwt.read. See 'dwt', and 'segment' for more information.}
\usage{
scan.read(filename, separating_factor = 1000)
}
\arguments{
\item{filename, }{the file name to read from.}

\item{separating_factor}{In lieu of a known time between segments, 
seperate with a multple of the longest dwell.}
}
\value{
A list of recording segments from the scan file
}
\description{
Data is in seconds.
}
\examples{

infile <- system.file("extdata", "example1_scan.txt", package = "scbursts")
record <- scan.read(infile)
head(record)

}
