% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.closed_dwells}
\alias{segment.closed_dwells}
\title{Extract closed dwells}
\usage{
segment.closed_dwells(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
the closed dwells
}
\description{
Extract closed dwells
}
\examples{
\dontrun{
closed_dwells <- segment.closed_dwells(data)
}
}
