% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.conductance_states}
\alias{segment.conductance_states}
\title{Return a list of all the (sub)conductance states.}
\usage{
segment.conductance_states(segment)
}
\arguments{
\item{segment}{The dwells and states table}
}
\value{
a list of all the (sub)conductance states.
}
\description{
Return a list of all the (sub)conductance states.
}
\examples{

# It's more likely that you created states or dwells with some function
states  <-  c(0,    0.2,    0,    1,    0,  0.5,    0,  0.7,    0,    1)
dwells  <-  c(0.1,  1.1,  0.5,  0.2,  1.0,  1.1,  0.6,  1.1,  0.8,  1.1)
my_burst <- segment.create(states, dwells, seg=1, start_time=3.14159, name="example_segment")

segment.conductance_states(my_burst)

}
