% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.consecutives_to_dwells}
\alias{segment.consecutives_to_dwells}
\title{Collapses a segment into dwells with alternating conductance levels.}
\usage{
segment.consecutives_to_dwells(segment)
}
\arguments{
\item{segment}{The dwells and states table}
}
\value{
A modified copy of the original segment
}
\description{
Segments may contain consecutive dwells with the same conductance level. 
consecutives_to_dwells sums together all consecutive dwells with the same
conductance level. The result is a segment containing dwells that alternate
in conductance level (i.e. 1,0,1,0,1,...)
}
