% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.copy}
\alias{segment.copy}
\title{Copy a segment}
\usage{
segment.copy(segment)
}
\arguments{
\item{segment}{The segment to copy}
}
\value{
A duplicate identical content.
}
\description{
This is a low-level function, mostly for use internally by other functions. There aren't many reasons to use this.
}
