% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.count_closed}
\alias{segment.count_closed}
\title{Extract number of closed dwells}
\usage{
segment.count_closed(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
number of closed dwells
}
\description{
Extract number of closed dwells
}
\examples{
\dontrun{
segment.count_closed(data)
> 5
}
}
