% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.count_dwells}
\alias{segment.count_dwells}
\title{Extract number of dwells in segment}
\usage{
segment.count_dwells(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
number of dwells
}
\description{
Extract number of dwells in segment
}
\examples{
\dontrun{
segment.count_dwells(data)
> 5
}
}
