% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.count_open}
\alias{segment.count_open}
\title{Extract number of open dwells}
\usage{
segment.count_open(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
number of open dwells
}
\description{
Extract number of open dwells
}
\examples{
\dontrun{
segment.count_open(data)
> 5
}
}
