% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.create}
\alias{segment.create}
\title{Create object containing table data and metadata}
\usage{
segment.create(states, dwells, seg = 1, start_time = 0, name = "burst")
}
\arguments{
\item{states}{a vector of states in {0,1}}

\item{dwells}{a vector of dwell durations (same length as states)}

\item{seg}{The segment number. Defaults to 1}

\item{start_time}{When the dwells began. Defaults to 0}

\item{name}{Suffix-less version of the original filename. 60uM.dwt -> '60uM'}
}
\value{
The segment object: A dataframe with extra metadata.
}
\description{
The object can be used as a dataframe, and the
metadata can be accessed with the functions
}
\details{
- segment.seg
- segment.start_time
- segment.filename
}
\examples{
\dontrun{
segment <- segment.create(states, dwells, seg=12, start_time=0, name="60uMc")
segment.name(segment)
> "60uMc"
}
}
