% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.duration}
\alias{segment.duration}
\title{Get duration of a segment}
\usage{
segment.duration(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
the duration
}
\description{
Get duration of a segment
}
\examples{
segment.duration(data)
> 2.11
}
