% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.name}
\alias{segment.name}
\title{Extract name from segment}
\usage{
segment.name(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
Segment name (string)
}
\description{
Extract name from segment
}
\examples{
segment.name(data)
> "60uMc"
}
