% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.open_dwells}
\alias{segment.open_dwells}
\title{Extract open dwells}
\usage{
segment.open_dwells(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
the open dwells
}
\description{
Extract open dwells
}
\examples{
\dontrun{
open_dwells <- segment.open_dwells(data)
}
}
