% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.pclosed}
\alias{segment.pclosed}
\title{Calculate empirical P(Closed) of a segment}
\usage{
segment.pclosed(segment)
}
\arguments{
\item{segment}{The dwells and states table}
}
\value{
The ratio of closed time to total time
}
\description{
NOTE: Assuming that burst starts and ends with 1
}
\examples{
\dontrun{
segment.pclosed(segment)
> 0.78
}
}
