% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.seg}
\alias{segment.seg}
\title{Extract segment number from segment}
\usage{
segment.seg(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
Segment number (integer)
}
\description{
Extract segment number from segment
}
\examples{
segment.seg(data)
> 12
}
