% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.start_time}
\alias{segment.start_time}
\title{Extract start_time from segment}
\usage{
segment.start_time(segment)
}
\arguments{
\item{segment}{the segment object}
}
\value{
Segment start_time (float)
}
\description{
Extract start_time from segment
}
\examples{
segment.start_time(data)
> 17.123295
}
