% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.verify}
\alias{segment.verify}
\title{Segments should have a very specific shape, but recordings can
produce errors that make non-sensical segments. In particular, ones
contain multiple consecutive openings or closings, or end in closings.}
\usage{
segment.verify(segment)
}
\arguments{
\item{segment}{The dwells and states table}
}
\value{
True if a valid segment, False otherwise
}
\description{
This function detects whether a segment satisfies the constraint that
the segment states alternate between open and closed, and begin and end with a closing.
}
\examples{
\dontrun{
segment.verify(segment)
> False
}
}
