% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{uic.plot_hist}
\alias{uic.plot_hist}
\title{Read a .evt file to a table}
\usage{
uic.plot_hist(points, title = "")
}
\arguments{
\item{points}{The data to plot}

\item{title}{The title of the plot. Defaults to blank}
}
\description{
Read a .evt file to a table
}
\examples{

open_times   <- subset(table, states == 1)$dwells
closed_times <- subset(table, states == 0)$dwells

uic.plot_hist(open_times, "Open Times")
uic.plot_hist(closed_times, "Closed Times")

}
