% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.basename}
\alias{util.basename}
\title{Remove suffix and path from filename.}
\usage{
util.basename(filename)
}
\arguments{
\item{filename}{string to extract basename from}
}
\value{
Name with suffix and path removed
}
\description{
Remove suffix and path from filename.
}
\examples{
util.basename("bursts/60uM-2017-08-18-16-32/60uM-712.dwt")
}
