% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{write_chunks_to_file}
\alias{write_chunks_to_file}
\title{Write all the bursts to files. They will be placed in
a timestamped subfolder of bursts/}
\usage{
write_chunks_to_file(chunks, directory = "bursts")
}
\arguments{
\item{chunks}{list of dataframes corresponding to bursts}
}
\value{
Timestamp
}
\description{
All bursts will be written out as .dwt files, labeled
in temporal order.
}
\examples{
write_chunks_to_file(chunks)
}
