% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.defined_by_tcrit}
\alias{bursts.defined_by_tcrit}
\title{Split segment at long pauses, dividing the segment
into multiple -shorter- segments (which are the bursts),
Along with the interburst closings, which are referred to as "gaps".}
\usage{
bursts.defined_by_tcrit(segment, t_crit)
}
\arguments{
\item{segment}{Segment with $states and $dwells}

\item{t_crit}{Critical time (us) at which to divide bursts}
}
\value{
A pair (bursts,gaps), where the bursts are segments
starting and ending in 1 states (open dwell), and gaps is a
vector of 0s (closed dwells) which sit inbetween the bursts.
There will be n bursts and n-1 gaps.
}
\description{
Split segment at long pauses, dividing the segment
into multiple -shorter- segments (which are the bursts),
Along with the interburst closings, which are referred to as "gaps".
}
\examples{
\dontrun{
# Splitting segment into smaller segments 
bursts_and_gaps <- bursts.defined_by_tcrit(record_c , 0.1)

# Extract the two parts
bursts <- bursts_and_gaps$bursts
gaps <- bursts_and_gaps$gaps

# Note that you need two brackets to access list elements.
# Also note that each burst begins and ends with an open dwell
# The gaps seperating them are the elements of gaps.
head(bursts[[11]])
>     states      dwells
> 427      0 15.16625000
> 428      1  0.31105000
> 429      0  0.01289401
> 430      1  0.04823000
> 431      0  0.04160000
> 432      1  0.14415000
}
}
