% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.filter}
\alias{bursts.filter}
\title{From a list of bursts, extract those that interest you by
passing a selecting function. See the examples.}
\usage{
bursts.filter(chunks, func, one_file = FALSE)
}
\arguments{
\item{chunks}{The list of all bursts}

\item{func}{A function of a segment that returns either TRUE or FALSE}
}
\value{
A shorter list of bursts OR if one_file is passed one segment
with zeros where the other bursts might have been originally.
}
\description{
From a list of bursts, extract those that interest you by
passing a selecting function. See the examples.
}
\examples{

high_popen <- function (seg) {

    segment.popen(seg) > 0.7

}

length(chunks)
> 733

subset <- bursts.filter(high_popen, chunks)

length(subset)
> 721


# To export to one .dwt file
subset_f <- bursts.filter(high_popen, chunks, one_file=TRUE)


}
