% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.get_gaps}
\alias{bursts.get_gaps}
\title{Extract vector of gaps from the bursts}
\usage{
bursts.get_gaps(bursts)
}
\arguments{
\item{bursts}{The list of records}
}
\value{
A vector of gap times
}
\description{
This is done using the start_time attribute, which
is mostly hidden in the data.
}
\examples{
\dontrun{
gaps <- bursts.get_gaps(bursts)
}
}
