% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.popens}
\alias{bursts.popens}
\title{Return popens of every burst.}
\usage{
bursts.popens(bursts)
}
\arguments{
\item{bursts}{The list of all bursts}
}
\value{
The popen values
}
\description{
Return popens of every burst.
}
\examples{
\dontrun{
popens <- bursts.popens(bursts)
hist(popens)
}
}
