% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.separate_tcrit}
\alias{bursts.separate_tcrit}
\title{Split record at long pauses, dividing the record
into multiple -shorter- bursts.}
\usage{
bursts.separate_tcrit(segment, t_crit)
}
\arguments{
\item{segment}{Segment with $states and $dwells}

\item{t_crit}{Critical time (us) at which to divide bursts}
}
\value{
A pair (chunks,breaks), where the chunks are segments
starting and ending in 1 states, and breaks is a vector of 0s which sit
inbetween the bursts. There will be n chunks and n-1 breaks.
}
\description{
Split record at long pauses, dividing the record
into multiple -shorter- bursts.
}
\examples{
pair <- dwt.read_bursts("bursts/60uM-2017-08-19-19-35")
chunks <- pair$chunks
breaks <- pair$breaks

# Note that lists are accessed with [[i]], not [i].

head(chunks[[11]])
>     states      dwells
> 427      0 15.16625000
> 428      1  0.31105000
> 429      0  0.01289401
> 430      1  0.04823000
> 431      0  0.04160000
> 432      1  0.14415000

}
