% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.start_times_update}
\alias{bursts.start_times_update}
\title{Attach the meta-data to each segment saying when it began.}
\usage{
bursts.start_times_update(chunks, breaks)
}
\arguments{
\item{chunks}{List of segments}

\item{breaks}{vector of break times.}
}
\value{
A list of segments, one per burst, with updated start_times
}
\description{
It interleaves the durations of the chunks and gaps, and
assigns the sum of those durations up to a point as the
starting time
}
\details{
You probably won't ever have to call this directly.
}
