% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.start_times_update}
\alias{bursts.start_times_update}
\title{(DON'T USE THIS) Fix meta-data of bursts.}
\usage{
bursts.start_times_update(bursts, gaps)
}
\arguments{
\item{bursts}{List of segments}

\item{gaps}{vector of gap times.}
}
\value{
A list of segments, one per burst, with updated start_times
}
\description{
YOU PROBABLY WON'T EVER HAVE TO CALL THIS DIRECTLY. Attach the meta-data to each segment saying when it began. It interleaves the durations of the bursts and gaps, and assigns the sum of those durations up to a point as the starting time.
}
