% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunk_bursts}
\alias{chunk_bursts}
\title{Split record at long pauses, dividing the record
into multiple -shorter- bursts.}
\usage{
chunk_bursts(table, t_crit)
}
\arguments{
\item{table}{Table with $states and $dwells}

\item{t_crit}{Critical time (us) at which to divide bursts}
}
\value{
A FUNCTION!!! chunk that returns the chunk by its index
}
\description{
Split record at long pauses, dividing the record
into multiple -shorter- bursts.
}
\examples{
chunk <- chunk_bursts(table, 14.77155587)
head(chunk(1))
>     states      dwells
> 427      0 15.16625000
> 428      1  0.31105000
> 429      0  0.01289401
> 430      1  0.04823000
> 431      0  0.04160000
> 432      1  0.14415000

}
