% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.log_root_axes}
\alias{cplot.log_root_axes}
\title{Add log-root axes to histogram plot}
\usage{
cplot.log_root_axes(points)
}
\arguments{
\item{points}{The data to plot}
}
\description{
Add log-root axes to histogram plot
}
\examples{

\dontrun{
hist(open_times, axes=FALSE)
cplot.log_root_axes(open_times)
}

}
