% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.log_root_hist}
\alias{cplot.log_root_hist}
\title{Plot points as a histogram after log transforming}
\usage{
cplot.log_root_hist(points, title = "")
}
\arguments{
\item{points}{The data to plot}

\item{title}{The title of the plot. Defaults to blank}
}
\description{
Plot points as a histogram after log transforming
}
\examples{

open_times   <- subset(table, states == 1)$dwells
closed_times <- subset(table, states == 0)$dwells

cplot.log_root_hist(open_times, "Open Times")
cplot.log_root_hist(closed_times, "Closed Times")

}
