% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.pclosed_ts}
\alias{cplot.pclosed_ts}
\title{Plot Time Series of P(Closed)}
\usage{
cplot.pclosed_ts(chunks, title = "P(Closed) Time Series", xlim = NULL)
}
\arguments{
\item{chunks}{List of multiple segments}

\item{title}{The title of the plot.}
}
\description{
Plot Time Series of P(Closed)
}
\examples{

\dontrun{
cplot.pclosed_ts(chunks, "P(Closed) Time Series, 2017-09-14")
}

}
