% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot.popen_ts}
\alias{cplot.popen_ts}
\title{Plot Time Series (ts) of P(Open)}
\usage{
cplot.popen_ts(bursts, main = "P(Open) Time Series", xlim = NULL)
}
\arguments{
\item{bursts}{List of multiple segments}

\item{main}{The title of the plot.}

\item{xlim}{Pass xlim argument to plot() to focus on window of time series}
}
\description{
Plot Time Series (ts) of P(Open)
}
\examples{

infile <- system.file("extdata", "example_corrected.dwt", package = "scbursts")
dwells_c <- dwt.read(infile)
bursts <- bursts.defined_by_tcrit(dwells_c, 1.511842, units="s")

cplot.popen_ts(bursts, "P(Open) Time Series, 2018-09-20")

}
