% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cplot.popen_ts}
\alias{cplot.popen_ts}
\title{Plot Time Series (ts) of P(Open)}
\usage{
cplot.popen_ts(bursts, main = "P(Open) Time Series", xlim = NULL)
}
\arguments{
\item{bursts}{List of multiple records}

\item{main}{The title of the plot.}
}
\description{
Plot Time Series (ts) of P(Open)
}
\examples{

\dontrun{
cplot.popen_ts(bursts, "P(Open) Time Series, 2017-09-14")
}

}
