% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.read}
\alias{dwt.read}
\title{Read a .dwt file. Result is a list of "segments", which is a
dataframe extra data. See "segment" for more details.}
\usage{
dwt.read(filename, separating_factor = 1000)
}
\arguments{
\item{filename}{Filename to read from}

\item{separating_factor}{In lieu of a known time between segments, seperate with a multple of the longest dwell.}
}
\value{
A list of bursts (possibly a singleton)
}
\description{
Converts millisecond dwells to seconds.
}
\examples{
\dontrun{
seg <- dwt.read(segment, "file-test.dwt")
}
}
