% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.read_bursts}
\alias{dwt.read_bursts}
\title{Read bursts from a folder.}
\usage{
dwt.read_bursts(folder)
}
\arguments{
\item{folder}{list of dataframes corresponding to bursts}
}
\value{
A pair (bursts,gaps), where the bursts are records
starting and ending in 1 states, and gaps is a vector of 0s which sit
inbetween the bursts. There will be n bursts and n-1 gaps.
}
\description{
Read bursts from a folder.
}
\examples{
\dontrun{
pair <- dwt.read_bursts("bursts/60uM-2017-08-19-19-35")
bursts <- pair$bursts
gaps <- pair$gaps
}
}
