% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.write}
\alias{dwt.write}
\title{Reformat the digits of data before writing to disk.
Should ask about this. At the moment this has a hard-coded
accuracy, because I don't understand significant figures.}
\usage{
dwt.write(record, file = "", rec = 1)
}
\arguments{
\item{record}{record with $dwells and $states}

\item{file}{Filename to write to}

\item{rec}{Record to write in .dwt header.}
}
\description{
Writes DOS line endings.
}
\examples{
\dontrun{
dwt.write(record, "file-test.dwt")
}
}
