% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.write}
\alias{dwt.write}
\title{Reformat the digits of data before writing to disk.}
\usage{
dwt.write(segments, file = "", seg = 1, append = FALSE)
}
\arguments{
\item{segments}{A segment or multiple segments with $dwells and $states}

\item{file}{Filename to write to}

\item{seg}{Segment number to write in .dwt header.}

\item{append}{Add ot the end of a file or overwrite? (defaults to false)}
}
\description{
Writes DOS line endings.
}
\details{
Dwells are in milliseconds
}
\examples{

infile <- system.file("extdata", "example.evt", package = "scbursts")
transitions <- evt.read(infile)
dwells <- evt.to_dwells(transitions)

dwt.write(dwells, file="dwells.dwt")

}
