% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt.write_bursts}
\alias{dwt.write_bursts}
\title{Write all the bursts to dwt. They will be placed in
a timestamped subfolder of bursts/}
\usage{
dwt.write_bursts(bursts, gaps = NULL, directory = "bursts",
  filename = "burst", timestamp = TRUE)
}
\arguments{
\item{bursts}{list of dataframes corresponding to bursts}

\item{gaps}{vector of gaps to write as csv (default is to try to reconstruct from bursts)}

\item{directory}{folder to create for bursts}

\item{filename}{label for the files}

\item{timestamp}{Put the date and time in the folder-name}
}
\value{
Name of folder containing contents.
}
\description{
All bursts will be written out as .dwt files, labeled
in temporal order.
}
\examples{
\dontrun{
dwt.write_bursts(bursts)
}
}
