% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.from_dwells}
\alias{evt.from_dwells}
\title{Converts dwell durations to absolute transition times.}
\usage{
evt.from_dwells(segments)
}
\arguments{
\item{segments}{A segment or multiple segemtns}
}
\value{
A dataframe or multiple dataframes of states and transition times
}
\description{
Converts dwell durations to absolute transition times.
}
\examples{

dwells_file <- system.file("extdata", "example1_qub.dwt", package = "scbursts")
dwells <- dwt.read(dwells_file)

transitions <- evt.from_dwells(dwells)

}
