% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.read}
\alias{evt.read}
\title{Read a .evt file to a table}
\usage{
evt.read(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
A list of tables with columns "states" and "times".
Each table corresponds to a contiguous segment from a recording.
}
\description{
Times are in seconds
}
\examples{

# import some of the data included with the package
infile <- system.file("extdata", "example.evt", package = "scbursts")
transitions <- evt.read(infile)

head(transitions[[1]])

}
