% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt.R
\name{evt.write}
\alias{evt.write}
\title{Write bursts to a .evt file}
\usage{
evt.write(segments, file = "", header = NULL)
}
\arguments{
\item{segments}{A segment or list of segments to write to filename}

\item{filename}{The filename}
}
\description{
Write bursts to a .evt file
}
\examples{
\dontrun{

evt.write(segments, file="60uMc-R.evt")

}
}
