% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst.R
\name{popen}
\alias{popen}
\title{Calculate empirical P(Open) of a *burst*}
\usage{
popen(table)
}
\arguments{
\item{table}{The dwells and states table}
}
\value{
The ratio of open time to total time
}
\description{
NOTE: Removing first and last state, as they are pauses.
}
\examples{
popen(table)
> 0.22
}
