% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read.evt}
\alias{read.evt}
\title{Read a .evt file to a table}
\usage{
read.evt(infile)
}
\arguments{
\item{infile}{The filename}
}
\value{
A table with columns "states" and "times"
}
\description{
Read a .evt file to a table
}
\examples{
table <- read.evt("data/60uM.evt")
}
