% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.popen}
\alias{segment.popen}
\title{Calculate empirical P(Open) of a segment}
\usage{
segment.popen(segment)
}
\arguments{
\item{segment}{The dwells and states table}
}
\value{
The ratio of open time to total time
}
\description{
NOTE: Assuming that burst starts and ends with 1
}
\examples{
\dontrun{
segment.popen(segment)
> 0.22
}
}
