% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{write.dwt}
\alias{write.dwt}
\title{Reformat the digits of data before writing to disk.
Should ask about this. At the moment this has a hard-coded
accuracy, because I don't understand significant figures.}
\usage{
write.dwt(table, filename, segment = 1)
}
\arguments{
\item{table}{Table with $dwells and $states}

\item{filename}{Filename to write to}
}
\description{
Writes DOS line endings.
}
\examples{
write.dwt(table, "file-test.dwt")
}
